[edit]: refactoring after id type changes

+++ src/by/epamtc/bakulin/dao/DAO.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface DAO<T> {
     void add(T entity) throws DAOException;
 
-    T findById(Long id) throws DAOException;
+    T findById(Integer id) throws DAOException;
 
     List<T> findAll() throws DAOException;
 
     void update(T entity) throws DAOException;
 
-    void delete(Long id) throws DAOException;
+    void delete(Integer id) throws DAOException;
 }
